Display All User's News Records
The application can now display all records by all users. Each post can display the correct title, url with <a> to access the url, date and time when the post was added to the table and the correct user's email and category text details. The post being displayed now has default values for upvote and downvote, the default value is 0.